Supervisor: use inet_http_server instead of socket
This change provides a different configuration for supervisord which
enables the inet_http_server instead of the one based on unix sockets.
The reason is that unix sockets are incompatible with gVisor sandboxed
containers.

Images based on this one and container configurations must make sure
that the given port (9091) is not exposed (even though the inet http
server binds to 127.0.0.1 only). If supervisorctl support is not
needed, images may want to remove `supervisorctl.conf` altogether.

Addressing flownative/beach#429