Reverting supervisorctl to unix_http_server
When multiple containers with supervisor are running in the same pod,
there will be port conflicts, as all containers of a pod share ports
on 127.0.0.1 and port 9091 would be used by the first supervisord
being started. A possible fix would be to make the port configurable,
however, this is error-prone and since we can't use gVisor sandboxes
at the moment anyway, we get back using the more secure default based
on unix sockets.

Note this change in flownative/beach#429