Add workflow for both, bullseye and buster
Scheduled Github workflow events can only be processed by the default
branch, so we need to use one branch for building both nightly builds,
for bullseye and for buster.

Therefore, I chose to declare bullseye as the new default branch and
moved the build workflow from buster to this branch.